Transform to IDW coordinate & fixed building int.parse to float.parse

+++ Assets/Scripts/Procedural/HouseGenerator.cs
@@ -30,7 +30,7 @@ static public class HouseGenerator
     {
         //demo code
         string[] house_infos = info.Split(' ');
-        Vector3 single_point = new Vector3(int.Parse(house_infos[2]), int.Parse(house_infos[3]), int.Parse(house_infos[4]));
+        Vector3 single_point = new Vector3(float.Parse(house_infos[2]), float.Parse(house_infos[3]), float.Parse(house_infos[4]));
         GameObject gobj = builder.build(component_name);
         gobj.transform.position = single_point;
         if (!gobj_db.ContainsKey(segment_id)) {
